Added new feature - NotificationUtils
NotificationUtils allows to subscribe for email notifications for
changes to an object's fields. Instead of enabling tracking history for
identifying the field changes, such changes would be included in the
email. You can choose the se of fields to be used to identify the
changes. In other words, fields to monitor for changes. Additionally,
you can configure the set of fields that holds the email address of
recipients